+++ src/renderer/viz/expressions/Image.js
@@ -28,7 +28,6 @@ export default class Image extends Base {
     constructor (url) {
         checkMaxArguments(arguments, 1, 'image');
         checkString('image', 'url', 0, url);
-
         super({});
         this.type = 'image';
         this.canvas = null;
@@ -58,8 +57,22 @@ export default class Image extends Base {
         }
     }
 
-    eval () {
+    get value () {
         return this.url;
+    }
+
+    eval () {
+        return this.value;
+    }
+
+    getLegendData (options) {
+        return {
+            name: 'image',
+            data: [{
+                key: 'url',
+                value: this.url
+            }]
+        };
     }
 
     _free (gl) {